requirements.txt
pyserial==3.4
pyqt==5.9.2
pyqtgraph==0.11.0
pandas==1.1.3
matplotlib==3.3.2
numpy==1.19.1